test: adjust case
