+++ test/cases/22-Functions/03-Selection/test_select_first_last.py
@@ -457,7 +457,7 @@ class TestSelectFirstLast:
             f"insert into tb3 values (\"2021-05-09 10:12:17\", 7, 8.0, '9' , -1002)"
         )
         tdSql.execute(
-            f'insert into tb3 values ("2021-05-09 10:12:17",10,11.0, NULL, -2002)'
+            f'insert into tb3 values ("2021-05-09 10:12:16",10,11.0, NULL, -2002)'
         )
         tdSql.execute(
             f'insert into tb3 values ("2021-05-09 10:12:18",12,NULL, NULL, -3002)'
@@ -503,14 +503,13 @@ class TestSelectFirstLast:
         tdSql.execute(f"alter table st2 add column c1 int;")
         tdSql.execute(f"alter table st2 drop column c1;")
 
-        self.last_both_query()
-
         tdSql.execute(f"flush database {db}")
         sc.dnodeStop(1)
         sc.dnodeStart(1)
         clusterComCheck.checkDnodes(1)
 
         self.last_both_query()
+        self.last_with_cols("last_with_cols")
 
         sc.dnodeStop(1)
         sc.dnodeStart(1)
@@ -596,7 +595,7 @@ class TestSelectFirstLast:
             f"insert into tb3 values (\"2021-05-09 10:12:17\", 7, 8.0, '9' , -1002)"
         )
         tdSql.execute(
-            f'insert into tb3 values ("2021-05-09 10:12:17",10,11.0, NULL, -2002)'
+            f'insert into tb3 values ("2021-05-09 10:12:16",10,11.0, NULL, -2002)'
         )
         tdSql.execute(
             f'insert into tb3 values ("2021-05-09 10:12:18",12,NULL, NULL, -3002)'
@@ -646,6 +645,7 @@ class TestSelectFirstLast:
 
         self.last_both_no_ts()
         self.last_both_query()
+        self.last_with_cols("last_with_cols")
 
     def last_both_query(self):
         db = "testdb"
@@ -709,8 +709,24 @@ class TestSelectFirstLast:
         tdSql.checkData(0, 3, None)
         tdSql.checkData(0, 4, None)
 
+        tdSql.query(f"select last(f1) from tbd")
+        tdSql.checkRows(0)
+
+        tdSql.query(f"select cols(last(f2), ts) from tbd")
+        tdSql.checkRows(0)
+
+        tdSql.query(f"select last(ts) from tbd")
+        tdSql.checkRows(1)
+
+        tdSql.query(f"select last(*) from tbd")
+        tdSql.checkRows(1)
+        tdSql.checkData(0, 0, "2021-05-11 10:12:29")
+
         tdLog.info(f'"test tbe"')
         tdSql.query(f"select last(*) from tbe")
+        tdSql.checkRows(0)
+
+        tdSql.query(f"select last_row(*) from tbe")
         tdSql.checkRows(0)
 
         tdLog.info(f'"test stable"')
@@ -755,7 +771,7 @@ class TestSelectFirstLast:
         tdSql.checkData(1, 5, 2)
         tdSql.checkData(2, 0, "2021-05-10 10:12:24")
         tdSql.checkData(2, 1, 24)
-        tdSql.checkData(2, 2, 11.000000000)
+        tdSql.checkData(2, 2, 8.000000000)
         tdSql.checkData(2, 3, 25)
         tdSql.checkData(2, 5, 3)
         tdSql.checkData(3, 0, "2021-05-11 10:12:25")
@@ -793,22 +809,9 @@ class TestSelectFirstLast:
         tdSql.checkData(3, 3, 27)
         tdSql.checkData(3, 5, 4)
         tdSql.checkData(4, 0, "2021-05-11 10:12:29")
-
-        # if $tdSql.getData(4,1, "NULL then
-        #  return -1
-        # endi
-        # if $tdSql.getData(4,2, "NULL then
-        #  print $tdSql.getData(0,2)
-        #  return -1
-        # endi
         tdSql.checkData(4, 3, None)
-
-        # if $tdSql.getData(4,4, "NULL then
-        #  return -1
-        # endi
         tdSql.checkData(4, 5, 5)
 
-        tdLog.info(f'"test tbn"')
         tdSql.execute(
             f"create table if not exists tbn (ts timestamp, f1 int, f2 double, f3 binary(10), f4 timestamp)"
         )
@@ -870,6 +873,12 @@ class TestSelectFirstLast:
         tdSql.checkData(0, 2, 37.000000000)
         tdSql.checkData(0, 3, 27)
         tdSql.checkData(0, 4, "1970-01-01 07:59:57")
+
+    def last_with_cols(self, testCase):
+        self.sqlFile = etool.curFile(__file__, f"in/{testCase}.in")
+        self.ansFile = etool.curFile(__file__, f"ans/{testCase}.csv")
+
+        tdCom.compare_testcase_result(self.sqlFile, self.ansFile, testCase)
 
     def ParserLastCache(self):
         tdLog.info(f"======================== dnode1 start")
