+++ test/ci/cases.task
@@ -261,7 +261,6 @@
 ,,y,.,./ci/pytest.sh pytest cases/22-Functions/03-Selection/test_select_top_bottom.py
 ,,y,.,./ci/pytest.sh pytest cases/22-Functions/03-Selection/test_select_with_520_cols.py
 ,,y,.,./ci/pytest.sh pytest cases/22-Functions/03-Selection/test_select_with_json.py
-,,y,.,./ci/pytest.sh pytest cases/22-Functions/03-Selection/test_func_last_both.py
 ## 04-Timeseries
 ,,y,.,./ci/pytest.sh pytest cases/22-Functions/04-Timeseries/test_ts_diff.py
 ## 05-Others
